+++ src/main/java/com/desk8432/project/controller/UpdatePassword.java
@@ -1,8 +1,6 @@
 package com.desk8432.project.controller;
 
-import com.desk8432.project.dao.UpdateEmailDAO;
-import com.desk8432.project.dao.UpdatePasswordDAO;
-import com.desk8432.project.dto.UpdateEmailDTO;
+import com.desk8432.project.dao.UpdateDAO;
 import com.desk8432.project.dto.UpdatePasswordDTO;
 import com.desk8432.project.util.Dispatcher;
 import com.google.gson.Gson;
@@ -26,11 +24,11 @@ public class UpdatePassword extends HttpServlet {
 
         Gson gson = new Gson();
         UpdatePasswordDTO updatePasswordDTO = gson.fromJson(jsonString, UpdatePasswordDTO.class);
-        UpdatePasswordDAO updatePasswordDAO = new UpdatePasswordDAO();
+        UpdateDAO updateDAO = new UpdateDAO();
 
         Gson outGson = new Gson();
         Map<String,String> resultMap = new HashMap<>();
-        if (updatePasswordDAO.updatePassword(updatePasswordDTO)) {
+        if (updateDAO.updatePassword(updatePasswordDTO)) {
             System.out.println("success");
             resultMap.put("message", "ok");
 //            ScriptWriter.alertAndNext(resp,"이메일 변경이 완료되었습니다","/index/index");
